fold calculateAspectRatioFit into image.utils
